# src/backend/docs/swagger.yaml
basePath: /
definitions:
  model.JSONErrorResult:
    properties:
      error: {}
      status:
        example: failed
        type: string
    type: object
  model.JSONSuccessResult:
    properties:
      data: {}
      status:
        example: success
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the back-end documentation of the pic-keeper project
  license:
    name: Apache 2.0
  title: Pic-keeper APIs
  version: "1.0"
paths:
...
  /authen/v1/google/callback:
    get:
      consumes:
      - application/json
      description: This will be automatically called when the Google OAuth2 login
        process is completed
      produces:
      - application/json
      responses:
        "200":
          description: The session token will be returned in the data field
          schema:
            allOf:
            - $ref: '#/definitions/model.JSONSuccessResult'
            - properties:
                data:
                  type: object
                status:
                  type: string
              type: object
        "500":
          description: Unhandled internal server error
          schema:
            allOf:
            - $ref: '#/definitions/model.JSONErrorResult'
            - properties:
                error:
                  type: object
                status:
                  type: string
              type: object
      summary: This will be automatically called when the Google OAuth2 login process
        is completed
      tags:
      - google
...
  /users/v1/upload-profile:
    post:
      consumes:
      - application/json
      description: Upload profile picture
      parameters:
      - description: Session token is required
        in: header
        name: Token
        required: true
        type: string
      - description: The profile picture file is required
        in: formData
        name: ProfilePicture
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Successfully uploaded the profile picture
          schema:
            allOf:
            - $ref: '#/definitions/model.JSONSuccessResult'
            - properties:
                data:
                  type: object
                status:
                  type: string
              type: object
        "400":
          description: Incorrect input
          schema:
            allOf:
            - $ref: '#/definitions/model.JSONErrorResult'
            - properties:
                error:
                  type: object
                status:
                  type: string
              type: object
        "500":
          description: Unhandled internal server error
          schema:
            allOf:
            - $ref: '#/definitions/model.JSONErrorResult'
            - properties:
                error:
                  type: object
                status:
                  type: string
              type: object
      summary: Upload profile picture
      tags:
      - users
swagger: "2.0"